+++ src/App.js
@@ -7,10 +7,10 @@ import {
 import { AnimatePresence } from "framer-motion";
 import Footer from "./components/Footer";
 import Nav from "./components/Nav";
-import Home from "./components/Home";
-import Profile from "./components/Profile";
-import Thanks from "./components/Thanks";
-import Todobubu from "./components/works/Todobubu";
+import Home from "./pages/home/Home";
+import Profile from "./pages/profile/Profile";
+import Thanks from "./pages/profile/Thanks";
+import Todobubu from "./pages/home/works/Todobubu";
 import "./App.css";
 
 const App = () => {
@@ -22,7 +22,7 @@ const App = () => {
       <AnimatePresence exitBeforeEnter>
         <Switch location={location} key={location.pathname}>
           <Route exact path="/" component={Home} />
-          <Route path="/profile" component={Profile} />
+          <Route exact path="/profile" component={Profile} />
           <Route path="/profile/thanks" component={Thanks} />
           <Route path={"/works/todobubu"} component={Todobubu} />
         </Switch>
