+++ src/assets/img/memoji.png
