+++ src/assets/img/thumb-todobubu.jpg
