+++ src/components/Work.js
@@ -1,6 +1,5 @@
 import React from "react";
 import { Link, Route, Switch, useRouteMatch } from "react-router-dom";
-import Todobubu from "./works/Todobubu";
 
 const Work = ({ src, wider, title, desc }) => {
   let { path, url } = useRouteMatch();
