+++ src/pages/home/Hero.js
